Extract body analysis result accumulator
Move metadata-ordered method result merging and assembly-level projections out of LibraryBodyAnalysisBuilder while preserving deterministic ordering and partial-result publication.
